# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Riyadh Metro Smart Mobility API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Station {
  id: string;
  name: string;
  nameAr: string;
  lineId: string;
  lineNumber: number;
  lineName: string;
  lineColor: string;
  isInterchange: boolean;
  interchangeLines: string[];
  order: number;
}

export interface MetroLine {
  id: string;
  number: number;
  name: string;
  nameAr: string;
  color: string;
  stations: Station[];
  totalStations: number;
}

export type StationDetailCrowdLevel = typeof StationDetailCrowdLevel[keyof typeof StationDetailCrowdLevel];


export const StationDetailCrowdLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export interface TrainArrival {
  direction: string;
  minutesUntil: number;
}

export interface StationDetail {
  id: string;
  name: string;
  nameAr: string;
  lineId: string;
  lineNumber: number;
  lineName: string;
  lineColor: string;
  isInterchange: boolean;
  interchangeLines: string[];
  order: number;
  crowdLevel: StationDetailCrowdLevel;
  nextTrains: TrainArrival[];
  amenities: string[];
}

export interface TripPlanInput {
  originId: string;
  destinationId: string;
  preferFewer?: boolean;
}

export type TripPlanCrowdLevel = typeof TripPlanCrowdLevel[keyof typeof TripPlanCrowdLevel];


export const TripPlanCrowdLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export interface TripStep {
  lineId: string;
  lineName: string;
  lineColor: string;
  lineNumber: number;
  boardAt: string;
  alightAt: string;
  stops: number;
  duration: number;
}

export interface TripPlan {
  origin: Station;
  destination: Station;
  /** Duration in minutes (simulated) */
  totalDuration: number;
  totalStops: number;
  /** Fare in SAR */
  fare: number;
  steps: TripStep[];
  crowdLevel: TripPlanCrowdLevel;
  /** Simulation note */
  note: string;
}

export interface SavedTripInput {
  originId: string;
  destinationId: string;
  originName: string;
  destinationName: string;
}

export interface SavedTrip {
  id: string;
  originId: string;
  destinationId: string;
  originName: string;
  destinationName: string;
  savedAt: string;
  userId: string;
}

export type NetworkSummaryAverageCrowdLevel = typeof NetworkSummaryAverageCrowdLevel[keyof typeof NetworkSummaryAverageCrowdLevel];


export const NetworkSummaryAverageCrowdLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export interface NetworkSummary {
  totalLines: number;
  totalStations: number;
  totalInterchanges: number;
  activeTrains: number;
  averageCrowdLevel: NetworkSummaryAverageCrowdLevel;
  operationalStatus: string;
}

export interface SignupInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: string;
  name: string;
  email: string;
  createdAt: string;
  rewardPoints?: number;
}