# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Riyadh Metro Smart Mobility API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  name: string;
  email: string;
  createdAt: string;
  rewardPoints?: number;
}