# lib/api-zod/src/generated/types/networkSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Riyadh Metro Smart Mobility API
 * OpenAPI spec version: 0.1.0
 */
import type { NetworkSummaryAverageCrowdLevel } from './networkSummaryAverageCrowdLevel';

export interface NetworkSummary {
  totalLines: number;
  totalStations: number;
  totalInterchanges: number;
  activeTrains: number;
  averageCrowdLevel: NetworkSummaryAverageCrowdLevel;
  operationalStatus: string;
}