# lib/api-zod/src/generated/types/savedTrip.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Riyadh Metro Smart Mobility API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedTrip {
  id: string;
  originId: string;
  destinationId: string;
  originName: string;
  destinationName: string;
  savedAt: string;
  userId: string;
}