# lib/api-zod/src/generated/types/savedTripInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Riyadh Metro Smart Mobility API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedTripInput {
  originId: string;
  destinationId: string;
  originName: string;
  destinationName: string;
}