# lib/api-zod/src/generated/types/signupInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Riyadh Metro Smart Mobility API
 * OpenAPI spec version: 0.1.0
 */

export interface SignupInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}