# lib/api-zod/src/generated/types/station.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Riyadh Metro Smart Mobility API
 * OpenAPI spec version: 0.1.0
 */

export interface Station {
  id: string;
  name: string;
  nameAr: string;
  lineId: string;
  lineNumber: number;
  lineName: string;
  lineColor: string;
  isInterchange: boolean;
  interchangeLines: string[];
  order: number;
}