# lib/api-zod/src/generated/types/stationDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Riyadh Metro Smart Mobility API
 * OpenAPI spec version: 0.1.0
 */
import type { StationDetailCrowdLevel } from './stationDetailCrowdLevel';
import type { TrainArrival } from './trainArrival';

export interface StationDetail {
  id: string;
  name: string;
  nameAr: string;
  lineId: string;
  lineNumber: number;
  lineName: string;
  lineColor: string;
  isInterchange: boolean;
  interchangeLines: string[];
  order: number;
  crowdLevel: StationDetailCrowdLevel;
  nextTrains: TrainArrival[];
  amenities: string[];
}