# lib/api-zod/src/generated/types/tripPlan.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Riyadh Metro Smart Mobility API
 * OpenAPI spec version: 0.1.0
 */
import type { Station } from './station';
import type { TripPlanCrowdLevel } from './tripPlanCrowdLevel';
import type { TripStep } from './tripStep';

export interface TripPlan {
  origin: Station;
  destination: Station;
  /** Duration in minutes (simulated) */
  totalDuration: number;
  totalStops: number;
  /** Fare in SAR */
  fare: number;
  steps: TripStep[];
  crowdLevel: TripPlanCrowdLevel;
  /** Simulation note */
  note: string;
}