# lib/api-zod/src/generated/types/tripStep.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Riyadh Metro Smart Mobility API
 * OpenAPI spec version: 0.1.0
 */

export interface TripStep {
  lineId: string;
  lineName: string;
  lineColor: string;
  lineNumber: number;
  boardAt: string;
  alightAt: string;
  stops: number;
  duration: number;
}